refactor: add image with stratched bg component
