+++ src/sections/Introduction/WhoSection.tsx
@@ -2,7 +2,7 @@ import Section from '../../components/Section';
 import { Highlight } from '../../components/Typo';
 import Store1stImageS from '../../statics/store-1st-small.png';
 import Store1stImageL from '../../statics/store-1st-large.png';
-import ImagesByMediaQuery from '../../components/ImagesByMediaQuery';
+import ImagesByMediaQuery from '../../components/image/ImagesByMediaQuery';
 
 const WhoSection = () => {
   return (
