+++ src/sections/PayPremium.tsx
@@ -3,14 +3,9 @@ import KakaoTalk from '../statics/kakaotalk.png';
 import styled from 'styled-components';
 import Typo from '../components/Typo';
 import mediaQuery from '../styles/mediaQuery';
+import ImageWithStretchedBg from '../components/image/ImageWithStretchedBg';
 
 const ImageContainer = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-const KakaoTalkImage = styled.img`
-  width: auto;
-  height: 327px;
   margin: 24px 5px 20px;
 `;
 
@@ -39,7 +34,7 @@ const PayPremium = () => {
     <Section bgColor="bg2" paddingT={60} paddingB={40}>
       <Section.Title>귀하의 자녀가 학습을 결심하고 결제를 요청하고 있습니다.</Section.Title>
       <ImageContainer>
-        <KakaoTalkImage src={KakaoTalk} alt="프리미엄 결제 요청 알림톡"/>
+        <ImageWithStretchedBg src={KakaoTalk} alt="프리미엄 결제 요청 알림톡" height='327px'/>
       </ImageContainer>
       <ResponseButton onClick={() => alert('결제 성공')}>
         <Typo type="textMB">
