+++ src/sections/question2/Question2.tsx
@@ -5,34 +5,24 @@ import NaverPay from '../../statics/naver-pay.png';
 import StoreBenefitGraph from '../../statics/store_benefit_graph.png';
 import styled from 'styled-components';
 import mediaQuery from '../../styles/mediaQuery';
-import ImagesByMediaQuery from '../../components/ImagesByMediaQuery';
+import ImagesByMediaQuery from '../../components/image/ImagesByMediaQuery';
 import DifferentRefundS from '../../statics/different-refund-small.png';
 import DifferentRefundL from '../../statics/different-refund-large.png';
-
-const ImageContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  margin-top: 44px;
-`;
+import ImageWithStretchedBg from '../../components/image/ImageWithStretchedBg';
 
 const RefundGraphSection = styled.section`
   background-color: ${({ theme }) => theme.bg3};
   padding: 60px 44px 59px 38px;
-
-  ${mediaQuery.large} {
-    display: flex;
-    justify-content: center;
-    img{
-      width: auto;
-      height: 331px;
-    }
-  }
 `;
 
 const ResponsivePadding = styled.div`
-  ${mediaQuery.large}{
+  ${mediaQuery.large} {
     padding: 0 130px;
   }
+`;
+
+const CouponMargin = styled.div`
+  margin-top: 44px;
 `;
 
 const Question2 = () => {
@@ -48,16 +38,16 @@ const Question2 = () => {
         <Section.Subtitle>자발적 학습을 위한</Section.Subtitle>
         <Section.Title>동기부여가 다르다</Section.Title>
         <Section.Content>학습량에 따른 Point지급으로 동기부여</Section.Content>
-        <ImageContainer>
-          <img src={NaverPay} alt={'네이버페이 최대 45000원 환급'} width="272px"/>
-        </ImageContainer>
+        <CouponMargin>
+          <ImageWithStretchedBg src={NaverPay} alt="네이버페이 최대 45000원 환급" height="199px"/>
+        </CouponMargin>
       </Section>
       <RefundGraphSection>
-        <img src={StoreBenefitGraph} width="100%" alt='공부할 수록 쌓이는 환급'/>
+        <ImageWithStretchedBg src={StoreBenefitGraph} alt="공부할 수록 쌓이는 환급" height="331px"/>
       </RefundGraphSection>
-      <Section bgColor='bg2' paddingT={60} paddingB={60}>
+      <Section bgColor="bg2" paddingT={60} paddingB={60}>
         <ResponsivePadding>
-          <ImagesByMediaQuery small={DifferentRefundS} large={DifferentRefundL} alt='타 서비스와 달리 공부하면 환급하는 수학대왕'/>
+          <ImagesByMediaQuery small={DifferentRefundS} large={DifferentRefundL} alt="타 서비스와 달리 공부하면 환급하는 수학대왕"/>
         </ResponsivePadding>
       </Section>
     </>
