+++ src/sections/question3/Question3.tsx
@@ -1,27 +1,17 @@
 import Section from '../../components/Section';
 import QuestionBox from '../../components/QuestionBox';
 import Typo, { Highlight } from '../../components/Typo';
-import ImagesByMediaQuery from '../../components/ImagesByMediaQuery';
+import ImagesByMediaQuery from '../../components/image/ImagesByMediaQuery';
 import AchievementS from '../../statics/achievement_small.png';
 import AchievementL from '../../statics/achievement_large.png';
 import styled from 'styled-components';
 import PositiveCycle from '../../statics/cycle.png';
-import mediaQuery from '../../styles/mediaQuery';
+import ImageWithStretchedBg from '../../components/image/ImageWithStretchedBg';
 
 const AchievementText = styled.h3`
   text-align: center;
   color: ${({ theme }) => theme.text2};
   margin-top: 60px;
-`;
-
-const CycleImage = styled.img`
-  width: 100%;
-  border-radius: 5px;
-
-  ${mediaQuery.large} {
-    width: auto;
-    height: 335px;
-  }
 `;
 
 const Question3 = () => {
@@ -52,7 +42,7 @@ const Question3 = () => {
           {'AI의 고도화로 더욱 빠른 성적 향상'}
         </Section.Content>
         <Section.Card>
-          <CycleImage src={PositiveCycle} alt="수학대왕 프리미엄의 선순환 구조"/>
+          <ImageWithStretchedBg src={PositiveCycle} alt="수학대왕 프리미엄의 선순환 구조" height="335px"/>
         </Section.Card>
       </Section>
     </>
